requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
decorator==4.4.2
distro==1.9.0
ffmpeg-python==0.2.0
filelock==3.18.0
Flask==3.1.1
fsspec==2025.5.1
future==1.0.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.7.0
moviepy==1.0.3
mpmath==1.3.0
narwhals==1.40.0
networkx==3.4.2
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
openai==1.82.0
openai-whisper==20240930
packaging==24.2
pandas==2.2.3
pillow==11.2.1
proglog==0.1.12
protobuf==6.31.0
pyarrow==20.0.0
pydantic==2.11.5
pydantic_core==2.33.2
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-vlc==3.0.21203
pytz==2025.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.25.1
setuptools==80.8.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.45.1
streamlit-autorefresh==1.0.1
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.9.0
toml==0.10.2
torch==2.7.0
tornado==6.5.1
tqdm==4.67.1
triton==3.3.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uuid==1.30
watchdog==6.0.0
Werkzeug==3.1.3
